% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{deaths}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 714 rows and 18 columns.}
\usage{
deaths
}
\description{
deaths dataset from Agrinst and Piscke (2015) "Mastering Metrics"
}
\details{
http://masteringmetrics.com/resources/
}
\keyword{datasets}
