% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetrendslm_main.R
\docType{package}
\name{fetrendslm-package}
\alias{fetrendslm}
\alias{fetrendslm-package}
\title{Fixed Effects linear models with trends}
\description{
Allows for partially out linear models with fixed effects and/or trends.
}
\author{
\strong{Maintainer}: Chandler Lutz \email{chandler.lutz@gmail.com}

}
