% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petersen}
\alias{petersen}
\title{petersen}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5000 rows and 4 columns.}
\usage{
petersen
}
\description{
Generate data from Petersen (2009; RFS)
}
\details{
https://www.kellogg.northwestern.edu/faculty/petersen/htm/papers/se/test_data.htm
}
\keyword{datasets}
